+++ library/src/main/java/com/vicidroid/amalia/ui/ViewDelegate.kt
@@ -1,0 +1,19 @@
+package com.vicidroid.amalia.ui
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
+import com.vicidroid.amalia.core.ViewEvent
+import com.vicidroid.amalia.core.ViewState
+
+interface ViewDelegate<S : ViewState, E : ViewEvent> {
+
+    val viewDelegateLifecycleOwner: LifecycleOwner
+
+    fun eventLiveData(): LiveData<E>? = null
+
+    /**
+     * Render a view state that is provided by the Presenter.
+     * The view delegate updates the UI accordingly.
+     */
+    fun renderViewState(state: S)
+}