+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Add lifecycle methods to view delegate (https://github.com/vicidroiddev/amalia/issues/4)
 * Enhance the concept of binding to a presenter (https://github.com/vicidroiddev/amalia/issues/5)
 * Pushstate should resolve the correct thread (https://github.com/vicidroiddev/amalia/issues/6)
+* Abstract ViewDelegate functionality to an interface https://github.com/vicidroiddev/amalia/issues/13